Moving getLibImage and getThemeImage into Theme class
